vcl: Extract test helpers
Most unit tests need to create some OpenCL data types like a context or
device.  A lot of tests we have so far don't have special requirements
for those, they just need a context/device to run one of the api
functions. That's why I created setup functions for some of those and
put them in a separate file.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>