vcl: Validate devices when creating a context
cb175626 iterates through devices to determine the correct value when an
invalid property is passed to clCreateContext(). This makes one of the asserts
in test_create_context fail, because it tests for invalid devices. Like this:

```
let invalid_device = 42;
assert_eq!(
    create_context(ptr::null(), 1, &invalid_device as *const _ as _, None, ptr::null_mut()),
    Err(CL_INVALID_DEVICE)
)
```

Since we don't currently check if a device is valid or not, it dereferences the
value for `invalid_device`. Since the value is 42 it creates an error:

```
test api::context::test::test_create_context
...
thread 'api::context::test::test_create_context'
panicked at
'misaligned pointer dereference: address must be a multiple of 0x8 but is 0x2a'
```

And that error shows up as a build error and thinks meson has a bug.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>