aco: Add a disassembler
Adds a disassembler to deal with a few issues we have with llvm:
- The disassembly may change between llvm versions which is annoying
  when writing tests.
- The process of reporting a bug with the llvm disassembler, that bug
  being fixed and the fix being part of a llvm release takes time.

The new disassembler uses a combination of hard coded parsing and tables
which is different to a solution like isaspec which can generate all the
code from an xml. The reason for this is that  isaspec cannot currently
deal with the level of complexity of the amd isa.

aco_disassembler.cpp contains core functionality like figuring out the
instruction format, extracting fields and printing operands, sdwa,
dpp... It also contains tables which are used by all generations and and
the disassembler for gfx6.

The aco_disassembler_gfx*.cpp files are there to avoid having one giant
aco_disassembler.cpp file. The handling of a certain generations can be
split between aco_disassembler.cpp as well as different generation
specific files if the encoding did not change too much.
disasm_{format}_gfx* helpers are only added if a generation changed the
encoding in such a way that it would require adding a lot of gfx_level
checks to the existing handling.