vcl: Add `clEnqueueNativeKernel`
`clEnqueueNativeKernel` is optional and only available for devices that
have `CL_EXEC_NATIVE_KERNEL` set in `CL_DEVICE_EXECUTION_CAPABILITIES`.
Otherwise the semantics are implementation defined.

For now, let's disable it to complete OpenCL 1.0. If it's required we
can come back to it and implement it later.

Since semantics are implementation defined it will return
`CL_INVALID_OPERATION`, because (luckily) that's already defined by the
spec to represent a device that cannot execute native kernels.

On top of that `CL_EXEC_NATIVE_KERNEL` is always overridden, regardless
of whether the device behind vcl supports it or not.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>