freedreno: don't encode build path into binaries
Encoding build-specific path into installed binaries is generally
frowned upon. It harms the reproducibility of the build and e.g.
OpenEmbedded now considers that to be an error.

Instead of hardcoding rnn_src_path into the RNN_DEF_PATH define specify
it manually when running the tests.

Additionaly make sure that RNN_PATH is set correctly when running `meson
devenv` in order to allow developers to call tools without installing
XML files to the target system.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>