radv/rt: Use direct dispatches for encoding on gfx6
Indirect unaligned dispatches don't work on gfx6, they dispatch
(x, y, z) workgroups, not threads. This means we dispatch 64x as much
work as we really need to. Instead, use a direct unaligned dispatch with
the worst-case internal node count, which should be closer to the amount
we really need.

Cc: mesa-stable