iris: ensure we don't end up with a dangling pointer
Coverity points out something that should not happen, but it can't
prove. When the const buffer and the input buffer are the same buffer,
that `pipe_resource_reference(cbuf->buffer, NULL)` won't free
cbuf->buffer, which would result in input->buffer being a dangling
pointer. This would be a bug in our reference counting, since the fact
that cbuf->buffer and input->buffer being equal should imply a reference
count of >= 2. This commit adds an assert to ensure that we don't end up
with a dangling pointer.