vulkan/wsi/wayland: Use frog-fifo-v1 protocol for FIFO
Use the frog-fifo-v1 protocol to implement fifo
under Wayland + Vulkan.

This fixes two major issues that significantly
impacts Wayland usability for both gaming and
desktop use right now:

1. Swapchain starvation caused by being required
to stall for the 'frame' event. This leads to it
only ever being possible to have 1 image in flight
at a time, which starves the GPU of work and can
lead to lower performance and stutters.

2. With this protocol, the compositor has more
information about the client using FIFO, and thus
can use an implementation defined virtual refresh
rate when the client's surface is not visible.

This is opposed to applications using FIFO
stalling completely and breaking functionality
when one of their surfacess is occluded.

Relevant KWin MR for support for this:
https://invent.kde.org/plasma/kwin/-/merge_requests/6474

Signed-off-by: Joshua Ashton <misyl@froggi.es>
Co-authored-by: Derek Foreman <derek.foreman@collabora.com>