tu: Add mechanism to mark an iova range for dumping
For DGC, we will emit PM4 commands to a preprocess buffer provided by
the user. This means that we should mark the preprocess buffer as
needing to be dumped, so that if the the GPU hangs in the middle of a
generated command buffer we can see it in crashdec. However the user
only provides a device address (i.e. iova) instead of a VkBuffer for the
preprocess buffer, whereas the kernel API expects us to mark the GEM BO.
Ideally we'd have a different kernel API, but for now add a mechanism to
map an iova back to a BO and use that to mark the iova range the user
passes in as needing to be dumped.