vcl: Use `clEnqueueNDRangeKernelMESA`
All three array pointers are optional, meaning: They can be NULL. But
the transport protocol uses the size of an array to determin if it
really is. If size is 0 the host will set the pointer to NULL, if it is
not 0 it will allocate an array.

This leads to a problem, because all three pointers use the same value
for their size. That's why it's possible to have size > 0 but pointer ==
NULL. In which case the host allocates an array, when it shouldn't.

The custom version of this functions use three additional parameters to
transport the size of each array individually. This involves some
redundancy, because the same size is transportet multiple times, but
resolves the problem, because each array can be checked individually.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>