meson: remove power8 optimization and deprecate the flag
GCC has removed the -mpower8-vector optimization, though the flag
remains so that existing builds don't break. GCC has considered that
this option (and the corresponding power9-vector optimization is a
workaround, and a poor option to use.

I have chosen to remove the optimization entirely, since it only applies
to Power8 and not more recent Power architectures. I also have no way to
test if any changes I make are effective, and don't want to risk turning
this into dead code. If someone can test and wants to send an update to
keep the optimization in place I'm happy to review that instead.

ref: https://gcc.gnu.org/pipermail/gcc-patches/2024-January/643454.html