meson: Allow finding Clang inside or outside the LLVM dir
For distros that ship Clang and LLVM in different prefixes (like NixOS)
our current approach will fail to find clang. It will look in the LLVM
directory, not find Clang, and give up. This is fixed by checking first
in the LLVM directory (the most common), and then outside if it was not
found.