nvk: do not submit empty pushes in nvk_queue_init_context_state
The only reason why this did not crash until now  is because NVKMD_ENGINE_3D
and NVKMD_ENGINE_COMPUTE are always set in the only queue we returned.

With the addition of a video queue, where no state must be submitted, the logic
would try to push an empty push_buffer. This is wrong and the first place where
it fails is when it tries to allocate a BO of size 0 for the push data.

Regardless of the current effort to get Vulkan Video supported in NVK, gating
nvk_queue_submit_simple on dw_count > 0 is always the right thiing to do.

Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>