intel/isl: Only set CMF on renderable views on Xe2+
The compression format is only used during rendering.

This prevents drivers from hitting an unreachable when we start enabling
CCS on linear surfaces which may have non-renderable and non-pow2 formats.

For now, use the surface format instead of the view format to look up
the CMF. This strategy returns the correct CMF for compressed surfaces.