WIP: tu: Support sparse residency for images
The tricky thing here is that we have to emulate the 64k "standard"
tile sizes in terms of the native 4k macrotiles. We do this by
manipulating which 4k pages get mapped, so that the mapping to implement
the tiling. Supporting the standard block sizes allows emulation layers
to claim D3D Tiled Resources Tier 2, which is required for the 12.0
feature level.