pan/bi: Add intrinsics to allow remapping color attachments dynamically
This is needed for VK_KHR_dynamic_rendering_local_read, which allows
shader side gl_FragDataX to be mapped to different color attachments.

What we do here is let the user pass a custom blend descriptor, likely
loaded from an FAU slot. Since the blend descriptor encodes the render
target, it allows for simple remapping that doesn't involve using a
if-ladder to write to the right output.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Reviewed-by: Mary Guillemard <mary.guillemard@collabora.com>