freedreno/rddecompiler: Write to an out_file instead of stdout directly
Future commits will allow splitting the produced C code into multiple
C sources to speed up compilation times in which case it wouldn't make
sense to output to stdout anymore.

Instead of writing directly to stdout, write to `out_file` instead,
so that later on we can easily change it to write to an actual file.

Signed-off-by: Karmjit Mahil <karmjit.mahil@igalia.com>