freedreno/decode/generate-rd: Build libgenerate_rd_resources
Running rddecompiler with `-m` splits out the code into multiple C
source files so build the resources into a `libgenerate_rd_resources`
static library to be linked with `generate_rd`.

`rddecompiler` produces an `input_resources.txt` file which is where
the user can manually add the sources to be built. Trying to do that
automatically with Meson seemed to be unreliable so we require this
extra step.

Signed-off-by: Karmjit Mahil <karmjit.mahil@igalia.com>