ir3/cp: translate movs->mov when possible
If a shared src gets folded into a movs, it will behave the same as a
mov. Make it a mov to allow further optimizations.

Note that maybe this should actually be done at the NIR level: whenever
the source of read_invocation is convergent, it becomes a no-op and can
be removed.

Signed-off-by: Job Noorman <jnoorman@igalia.com>