util/disk_cache: unify asprintf error handling
Do not check whether the first parameter is NULL, but use the return
value to check for a possible error.

This matches what the other asprintf calls in disk_cache_os.c are doing,
and it is argubly also the correct way to check for asprintf failure:
On at least Linux, the contents of the first argument are undefined if
memory allocation failure occurs, and the documentation mentions that
in that case, the return value can be used for checking [1].

This effectively reverts 9033e64b022cc2da84b51fa5b2aa2133f8a0e6a9.

1. https://www.man7.org/linux/man-pages/man3/asprintf.3.html

Reviewed-by: Timothy Arceri <tarceri@itsqueeze.com>