HACK: assorted nil_image hacks to make video work
This is a collection of hacks that we should figure out properly in the future.

For example, we hardcode y_log2=4 and force the this value across all planes in
a nvk_image, among other things.

y_log2=4 seems to be required by the video codec engine, and it has to be the
same for both the luma and chroma planes.

As I do not have the time to figure this out (together will all the plumbing to
make it happen nicely), leave this hack for the time being.