nvk: video: compute the size of the opaque buffers correctly
We currently have random constants that are completely unsuitable for anything
larger than 360p.

The values used must scale with the actual video parameters, otherwise the GPU
will trigger IOMMU faults that will kill the channel.

Thanks to Averne for figuring out the right math for this in their Tegra ffmpeg
work.

Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>