nvk: video: add h.264 decode support in Rust
The current code will decode a single single frame at best in the vast majority
of cases before failing with VK_ERROR_DEVICE_LOST.

This change implements video decode support for the h.264 codec in Rust and
brings the support to 123/135 in the JVT-AVC_V1 test suite.

For the sake of comparison, ffmpeg currently sits at 121/135 using the VA-API
hwaccel.

Thanks to Averne for their work on the tracer that made this
reverse-engineering work possible.

Signed-off-by: Daniel Almeida <daniel.almeida@collabora.com>