DRAFT: wsi: Skip render-display verification when `/dev/dri/renderD*` isn't available
The wsi code tries to verify if the render node is the same as the
display, in the scenario that we passthrough the Virglrender, wayland
and/or X sockets to the container but not the `/dev/dri/renderD*` node,
we hit an issue where as the display fd node isn't present, thw wsi
code fails.

This code skips the verification.

Signed-off-by: Igor Torrente <igor.torrente@collabora.com>