nouveau: add g_nv_name_override.h based on the NVIDIA 340.108 driver
Add device name mappings replicating the format of g_nv_name_released.h
to aid better display of card names in nouveau. Mappings were dumped
from the NVIDIA 340.108 driver README.txt file which provides the first
two colums of data so the third has been padded to 0x0000. Items are
listed in the original order of README.txt which is not sorted on the
first column.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>