vbo: deal with pipe_vertex_state immutability
pipe_vertex_state has the following constraint:
   "All buffer contents are considered immutable and read-only"

which conflicts with the idea of recycling buffers.
To work around this issue, this commit introduces a dirtyness
tracking bit. When this bit is set for a range, it means it may
have a inflight user.

To avoid using st_glFinish which would introduce large stalls,
a deferred fence is grab when a BO is released. This way, we
can wait on this fence only before reusing the BO.

Even if this fence is not tracking ranges separately, it seems
good enough and doesn't regress performance.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/28688>