panfrost: allow up to 4 spills per pass in register allocation
The bifrost register allocator is extremely slow. Speed it up
by performing the best 4 spills (rather than just the best one)
each iteration of the register allocation loop. This speeds up
the process roughly 4x while not changing the number of spills
inserted, at least on our shaderdb code and the large vulkan
test case that initially prompted this.

Increasing the number of spills collected is trivial, but worsens
the results in testing. A more sophisticated algorithm to figure out
the number of spills to do each pass might be useful. We're working
on a new register allocator, and calculating that might be a first step.