anv: improve accuracy of intel_measure ends
Previously, end timestamps were emitted just prior to the
beginning of the next event. This is not accurate as any flushes
or stalls will extending the duration of the previous event. It
would be better to have stall/flush durations be counted towards
pre-event times, reported in a separate column.

This commit does it a better way: instrument accurate endings
for all events instead the old, implicit ending. Flushes/stalls
should now appear in pre-event results column.