anv: Use pipeline topology for GS shader input in Clanfolk
The geometry shaders in this game specify input topology to be points, but
the real pipeline topology is triangles, which causes the calls to be
ignored, as per "3DSTATE_GS_BODY::Expected Vertex Count" description. This
is an application bug, because both Vulkan and DX spec say that the
pipeline topology and GS input topology must match.
To fix the misrendering this commit adds a workaround that just computes
the "Expected Vertex Count" value directly from the pipeline topology.

Cc: mesa-stable
Closes: https://gitlab.freedesktop.org/mesa/mesa/-/issues/11853
Signed-off-by: Sviatoslav Peleshko <sviatoslav.peleshko@globallogic.com>