brw: remove the Mesa-specific letter casing quirks in Gen assembly grammar
I was playing with src/intel/executor and trying to use an instruction
that would take an immediate value similar to 0x0FQ. I kept getting
syntax errors even though everything was correct. It took me a long
time to realize that Mesa Gen assembly has specific rules regarding
casing: hex letters from immediate values have to be written in
lower-case, while the types have to be written in upper-case, so what
I had to write was 0x0fQ instead. I couldn't find this information
documented anywhere, I had to debug this all the way to brw_lex.l.

This is a Mesa-specific issue. According to BSpec pages "Instruction
Syntax" (56717) and "Grammar" (56736), what separates the digits from
an immediate definition and its type is the ':' character. Also, the
types are defined as lower-case instead of upper-case (see RegType or
ImmType in Grammar). The values accepted can be both upper and lower
case. So, according to the spec, I could have written 0x0f:q or
0x0F:q.

So in this patch we make the following changes:

- We start printing the ':' character.
- We accept both upper-case and lower-case immediate values.
- We accept both upper-case and lower-case type definitions.

So we're still different from BSpec since the ':' is still optional
for us, and we accept upper-case type definitions, but at least we are
now printing the ':' character everywhere, so when people copy/paste
assembly, everything should still work.

While the ':' character is still optional, we won't parse 0xfF anymore
as decimal 15 of type Float, we'll start parsing it as decimal 255
without a type (i.e., a syntax error). This is a
non-backwards-compatible change. On the other hand, I truly believe
this is better whan failing to accept 0x0FQ and having the user guess
that hex digits should be lower-case while type definitions should be
upper-case.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>