nak: start adding Turing latency information.
This is just an example of how bad this is.

There are some categories I need to work out where they are mapped to,
I've no idea if this stuff is hooked into the core properly either.

v2: add pred macro (finish this), to make comparison to docs easier
v3: move to a separate file, finish using pred macro everywhere
v4: actually run CTS, this fails a bunch WIP