tu: use 64bit_global_32bit_offset for global addresses in NIR
Keep using 64bit_global at the ABI level but switch to
64bit_global_32bit_offset at the NIR level.

This saves quite a lot of 64bit math for inbounds accesses since the
offsets will be calculated in 32bit.

This also paves the way to enable full support for ldg.a/stg.a which
take a 32bit offset as input.