spirv: allow different NIR/ABI address formats
For some backends and some memory modes, it might be useful to use
different pointer representations at the ABI level (e.g., how things are
stored in buffers) and at the NIR level (e.g., how things are computed
with).

This commit implements this for SpvStorageClassPhysicalStorageBuffer.

See the next commit for an example of how this is useful in Turnip.

Signed-off-by: Job Noorman <jnoorman@igalia.com>